Add files via upload
First the Server class must be run, then the client in order to start chatting/communicating. The IP addresses also need to be changed to your custom IP addresses in order to make it run. In order to run the program you need a developers kit that can run java files.